// src/types/a2a.ts
  export interface Message {
    /** Message sender's role */
    role: "user" | "agent";
    /** Message content */
    parts: Part[];
    /** Extension metadata. */
    metadata?: {
      [key: string]: any;
    };
    /** List of tasks referenced as context by this message.*/
    referenceTaskIds?: string[];
    /** Identifier created by the message creator*/
    messageId: string;
    /** Identifier of task the message is related to */
    taskId?: string;
    /** The context the message is associated with */
    contextId?: string;
    /** Event type */
    kind: "message";
  }
  
  /** Base properties common to all message parts. */
  export interface PartBase {
    /** Optional metadata associated with the part. */
    metadata?: {
      [key: string]: any;
    };
  }
  
  /** Represents a text segment within parts.*/
  export interface TextPart extends PartBase {
    /** Part type - text for TextParts*/
    kind: "text";
    /** Text content */
    text: string;
  }
  
  /** Represents the base entity for FileParts */
  export interface FileBase {
    /** Optional name for the file */
    name?: string;
    /** Optional mimeType for the file */
    mimeType?: string;
  }
  
  /** Define the variant where 'bytes' is present and 'uri' is absent */
  export interface FileWithBytes extends FileBase {
    /** base64 encoded content of the file*/
    bytes: string;
    uri?: never;
  }
  
  /** Define the variant where 'uri' is present and 'bytes' is absent  */
  export interface FileWithUri extends FileBase {
    /** URL for the File content */
    uri: string;
    bytes?: never;
  }
  
  /** Represents a File segment within parts.*/
  export interface FilePart extends PartBase {
    /** Part type - file for FileParts */
    kind: "file";
    /** File content either as url or bytes */
    file: FileWithBytes | FileWithUri;
  }
  
  /** Represents a structured data segment within a message part. */
  export interface DataPart extends PartBase {
    /** Part type - data for DataParts */
    kind: "data";
    /** Structured data content 
    */
    data: {
      [key: string]: any;
    };
  }
  
  /** Represents a part of a message, which can be text, a file, or structured data. */
  export type Part = TextPart | FilePart | DataPart